requirements.txt
python-docx
PyPDF2
tqdm
jsonschema
google-genai
genai